SC
As per Goldbach's Conjecture, any even number more than two can be represented as the sum of two prime numbers. For example 8: (3, 5); (5,3). It stayed a "conjecture" because of it's difficulty to actually prove 

Run this Python script in your Editor/IDE and see every single combination in your cmd/terminal